Real.Time.Queue

Description

Abstracts the implementation details of a single-insertion, single-extraction queuelike structure with amortised complexity.

Documentation

data Queue a

Constructors

Queue [a] [a]

newQ :: Queue a

Constructs an empty queue.

empty :: Queue a -> Bool

Checks if the queue is empty.

push :: Queue a -> a -> Queue a

Inserts a single element into the Queue.

pop :: Queue a -> Queue a

Attempts to extract an element from the Queue. Queue must not be empty!

front :: Queue p -> p

Gets the element that will next be extracted from the Queue. Queue must not be empty!front :: Queue a -> a

rev :: Queue a -> Queue a

Returns the elements of the Queue in reverse order.

size :: Queue a -> Int

Gets the size of the queue